pyproject.toml
[project]
name = "infrastructure"
version = "0.1.0"
description = "Homelab Infrastructure-as-Code using Ansible"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ansible>=9.0.0",
    "ansible-lint>=24.0.0",
    "yamllint>=1.35.0",
    "molecule>=6.0.0",
    "molecule-plugins[delegated,docker]>=23.5.0",
    "docker>=7.0.0",
    "requests>=2.31.0",
    "jinja2>=3.1.0",
    "pyyaml>=6.0.0",
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.8.0",
    "pyright>=1.1.0",
    "pre-commit>=3.5.0",
]

[dependency-groups]
dev = [
    "ruff>=0.8.0",
    "pyright>=1.1.0",
    "pre-commit>=3.5.0",
]

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = []

